backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic[email]==2.9.2
bcrypt==4.2.0
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
alembic==1.16.2
python-multipart==0.0.12
